requirements.txt
langchain_openai
streamlit
python-dotenv
langchain_core
langchain_community
langserve
fastapi
uvicorn
sse_starlette
beautifulsoup4
pypdf
chromadb
faiss-cpu
wikipedia-api
arxiv
groq
cassio
langchain-groq
langchainhub
validators==0.28.1
youtube_transcript_api
unstructured
unstructured[local-inference]
validators
pytube